add privileg users

+++ admin/administrators.php
@@ -18,7 +18,14 @@ if ($_SESSION['auth_admin'] == "yes_auth")
         switch ($action)
         {
             case 'delete':
-            $delete = mysql_query("DELETE FROM reg_admin WHERE id = '$id'",$link);
+            if($_SESSION['auth_admin_login'] == 'admin')
+            {
+                $delete = mysql_query("DELETE FROM reg_admin WHERE id = '$id'",$link);
+            }
+            else
+            {
+                $msgerror = 'У Вас немає прав на видалення адміністраторів!!!';
+            }
             break; 
         }
     }   
@@ -45,7 +52,10 @@ if ($_SESSION['auth_admin'] == "yes_auth")
             <p id="title-page"><strong>Адміністратори</strong></p>
             <p align="right" id="add-style"><a href="add_administrators.php">Додати адміна</a></p>
         </div>
-        <?php 
+        <?php
+        if(isset($msgerror)) echo '<p id="form-error" align="center">'.$msgerror.'</p>'; 
+        if($_SESSION['view_admin'] == '1')
+        {
             $result = mysql_query("SELECT * FROM reg_admin ORDER BY id DESC", $link);
             if(mysql_num_rows($result)>0)
             {
@@ -67,6 +77,14 @@ if ($_SESSION['auth_admin'] == "yes_auth")
                 }
                 while($row = mysql_fetch_array($result));
             }
+        }
+        else
+        {
+            echo 
+            '
+                <p id="form-error" align="center">У Вас немає прав на перегляд адміністраторів!!!</p>
+            ';
+        }
         ?>
     </div>
     </div>
